+++ shared/github.ts
@@ -1,11 +1,4 @@
 import fetch from 'node-fetch';
-import { findInMongoDb } from './azure-cosmosdb-data-to-mongodb';
-
-type DbConfigSetting = {
-  ownerWithRepo: string;
-  action: string;
-  dispatchType: string;
-};
 
 export type EnvironmentVariables = {
   pat: string;
@@ -28,7 +21,7 @@ export type BuildDispatchParams = {
   ContextFunctions;
 
 export type TriggerDispatchParams = {
-  dispatchType: string;
+  type: string;
   owner: string;
   repo: string;
   pat: string;
@@ -39,58 +32,14 @@ export type DispatchActionResult = {
   statusCode: number;
 };
 
-export async function dispatchAction(
-  vars: BuildDispatchParams
-): Promise<DispatchActionResult> {
-  if (!vars.owner) throw Error('owner is missing');
-  if (!vars.repo) throw Error('repo is missing');
-  if (!vars.action) throw Error('action is missing');
-  if (!vars.pat) throw Error('pat is missing');
-  if (!vars.dbConnectionString) throw Error('dbConnectionString is missing');
-  if (!vars.databaseName) throw Error('databaseName is missing');
-  if (!vars.collectionName) throw Error('collectionName is missing');
-  if (!vars.partitionKey) throw Error('partitionKey is missing');
-
-  const query = {
-    [vars.partitionKey]: { $eq: `${vars.owner}/${vars.repo}` },
-    action: { $eq: vars.action }
-  };
-
-  const options = {};
-
-  // get config setting from database
-  const configSetting = await findInMongoDb(
-    vars.dbConnectionString,
-    vars.databaseName,
-    vars.collectionName,
-    query,
-    options
-  );
-
-  if (!configSetting || configSetting.length === 0)
-    throw Error('config setting not found');
-
-  const { dispatchType } = configSetting?.[0] as DbConfigSetting;
-
-  const { statusCode } = await triggerDispatch({
-    dispatchType, //'data-is-ready',
-    owner: vars.owner,
-    repo: vars.repo,
-    pat: vars.pat,
-    log: vars.log
-  });
-
-  return { statusCode };
-}
-
 export async function triggerDispatch({
-  dispatchType,
+  type,
   owner,
   repo,
   pat,
   log
 }: TriggerDispatchParams) {
-  if (!dispatchType) throw Error('dispatchType is missing');
+  if (!type) throw Error('dispatchType is missing');
   if (!owner) throw Error('owner is missing');
   if (!repo) throw Error('repo is missing');
   if (!pat) throw Error('pat is missing');
@@ -105,7 +54,7 @@ export async function triggerDispatch({
       Authorization: `token ${pat}`
     },
     body: JSON.stringify({
-      event_type: dispatchType
+      event_type: type
     })
   };
 
@@ -114,7 +63,7 @@ export async function triggerDispatch({
   const response = await fetch(url, options);
   if (response.ok) {
     // No data is returned - status is 204
-    log(`dispatch success: ${owner}/${repo} with type ${dispatchType} }`);
+    log(`dispatch success: ${owner}/${repo} with type ${type} }`);
     return { statusCode: response.status };
   } else {
     // 404 or 401